+++ demo/src/MemoryModelsUserInput.tsx
@@ -5,7 +5,6 @@ import {
     Button,
     Input,
     TextField,
-    Tooltip,
     MenuItem,
     Stack,
     Dialog,
@@ -15,7 +14,6 @@ import {
 import { useTranslation } from "react-i18next";
 import UploadFileIcon from "@mui/icons-material/UploadFile";
 import FolderOpenIcon from "@mui/icons-material/FolderOpen";
-import DrawIcon from "@mui/icons-material/Draw";
 import DownloadJSONButton from "./DownloadJSONButton.js";
 import MemoryModelsMenu from "./MemoryModelsMenu.js";
 import MemoryModelsSample from "./MemoryModelsSample.js";
@@ -35,6 +33,8 @@ type MemoryModelsFileInputPropTypes = {
     setTextData: React.Dispatch<React.SetStateAction<string>>;
     textData: string;
     setFailureBanner: React.Dispatch<React.SetStateAction<string>>;
+    queueDebouncedTextData: (textData: string) => void;
+    flushDebouncedTextData: () => void;
 };
 
 type MemoryModelsTextInputPropTypes = {
@@ -46,7 +46,6 @@ type MemoryModelsTextInputPropTypes = {
 type MemoryModelsUserInputPropTypes = MemoryModelsFileInputPropTypes &
     MemoryModelsTextInputPropTypes &
     MemoryModelsConfigInputPropTypes & {
-        onTextDataSubmit: (event?: React.MouseEvent<HTMLFormElement>) => void;
         failureBanner: string;
     };
 
@@ -81,6 +80,8 @@ function MemoryModelsFileInput(props: MemoryModelsFileInputPropTypes) {
     const onLoadButtonClick = () => {
         props.setTextData(uploadedFileString);
         setOpen(false);
+        props.queueDebouncedTextData(uploadedFileString);
+        props.flushDebouncedTextData();
     };
 
     return (
@@ -230,20 +231,22 @@ function MemoryModelsConfigInput(props: MemoryModelsConfigInputPropTypes) {
 export default function MemoryModelsUserInput(
     props: MemoryModelsUserInputPropTypes
 ) {
-    const { t } = useTranslation();
     return (
-        <form data-testid="input-form" onSubmit={props.onTextDataSubmit}>
+        <form data-testid="input-form">
             <Stack spacing={2}>
                 <MemoryModelsFileInput
                     textData={props.textData}
                     setTextData={props.setTextData}
                     setFailureBanner={props.setFailureBanner}
+                    queueDebouncedTextData={props.queueDebouncedTextData}
+                    flushDebouncedTextData={props.flushDebouncedTextData}
                 />
                 <Box sx={{ display: "flex", justifyContent: "space-between" }}>
                     <MemoryModelsSample
                         setTextData={props.setTextData}
                         setConfigData={props.setConfigData}
-                        onTextDataSubmit={props.onTextDataSubmit}
+                        queueDebouncedTextData={props.queueDebouncedTextData}
+                        flushDebouncedTextData={props.flushDebouncedTextData}
                     />
                     <MemoryModelsConfigInput
                         configData={props.configData}
@@ -270,21 +273,6 @@ export default function MemoryModelsUserInput(
                     sx={{ justifyContent: "space-between" }}
                 >
                     <DownloadJSONButton textData={props.textData} />
-                    <Tooltip title={t("input.drawTooltip")}>
-                        <span>
-                            <Button
-                                type="submit"
-                                data-testid="input-submit-button"
-                                variant="contained"
-                                color="primary"
-                                disabled={!props.textData}
-                                style={{ textTransform: "none" }}
-                                startIcon={<DrawIcon />}
-                            >
-                                {t("input.drawDiagram")}
-                            </Button>
-                        </span>
-                    </Tooltip>
                 </Stack>
             </Stack>
         </form>